Script: move WeBWorK processing into "webwork.py"
